% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cynkra-desc.R
\name{register_cynkra}
\alias{register_cynkra}
\alias{add_cynkra_ror}
\title{Register cynkra in \code{DESCRIPTION}}
\usage{
register_cynkra(role = c("fnd", "cph"))

add_cynkra_ror()
}
\arguments{
\item{role}{Role to add to cynkra. See \code{utils:::MARC_relator_db_codes_used_with_R}.}
}
\description{
Add cynkra metadata to the package

Add cynkra's ROR when cynkra is already an author.
}
