% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cynkrathis-package.R
\docType{package}
\name{cynkrathis-package}
\alias{cynkrathis}
\alias{cynkrathis-package}
\title{cynkrathis: Simplify Project Scaffolding and Common 'renv' Related Tasks}
\description{
Opionionated renv helper package by cynkra GmbH.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cynkrathis.cynkra.com/}
  \item \url{https://github.com/cynkra/cynkrathis}
  \item Report bugs at \url{https://github.com/cynkra/cynkrathis/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

Authors:
\itemize{
  \item Kirill Müller \email{krlmlr+r@mailbox.org} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item cynkra GmbH \email{mail@cynkra.com} (0335t7e62) [funder]
}

}
\keyword{internal}
