% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_snapshots}
\alias{get_snapshots}
\title{Return metadata for cynkra RSPM snapshots}
\usage{
get_snapshots()
}
\description{
A list of curated cynkra RStudio Package Manager snapshot IDs and their
 metadata.
}
