% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_valid_snapshots}
\alias{get_valid_snapshots}
\title{Return metadata for RSPM snapshots}
\usage{
get_valid_snapshots()
}
\description{
A list of valid snapshot IDs and their corresponding data and R version.
}
