% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{init_gitignore}
\alias{init_gitignore}
\title{Write fully equipped gitignore files}
\usage{
init_gitignore(commit = TRUE)
}
\arguments{
\item{commit}{\verb{[logical]}\cr
Whether to automatically commit the newly created files.}
}
\description{
By making use of the gitignore.io API.
}
