% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precommit.R
\name{init_precommit}
\alias{init_precommit}
\title{Use precommit with that project}
\usage{
init_precommit(commit = TRUE)
}
\arguments{
\item{commit}{`[logical]`\cr
Whether to automatically commit the newly created files.}
}
\description{
Use precommit with that project
}
