% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{init_renv}
\alias{init_renv}
\title{Initialize {renv} infrastructure (cynkra way)}
\usage{
init_renv(
  snapshot_date = NULL,
  additional_repos = NULL,
  local_packages = NULL,
  exclude_local = NULL
)
}
\arguments{
\item{snapshot_date}{\verb{[integer]}\cr
A valid RSPM snapshot date. By default the "recommended" date from
\code{get_snapshots()} for the respective R version is chosen.}

\item{additional_repos}{\verb{[named character]}\cr
Additional repos besides the RSPM one.
Must be a named character vector.}

\item{local_packages}{\verb{[list]}\cr
Packages to exclude from the \code{renv::install()} call and to install
from local sources.
Needs the package name and an explicit version.
This is useful for local package sources used in the project which are not
available in one of the repos configured in the project.}

\item{exclude_local}{\verb{[named character]}\cr
Local packages to exclude from \code{renv::install()}.
Required if the package is only available locally.
If \code{local_packages} was set, this argument can be ignored.}
}
\description{
Initializes {renv} setup by setting a predefined RStudio Package
Manager (RSPM) snapshot.
Optionally more 'minicran' repositories can be configured via argument
\code{additional_repos}.
Custom RSPM Snapshots can be configured via \code{snapshot_date}.
}
\details{
During the process, the latest CRAN version of {renv} will be installed,
regardless of the chose snapshot ID.
}
\examples{
\dontrun{
init_renv(
  additional_repos = c(e360 = "https://$DOMAIN/drat"),
  local_packages = structure(c(foo = "0.1.0", foo2 = "0.2.1"))
)
}
}
