% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.R
\name{init_tic}
\alias{init_tic}
\title{Initialize tic CI structure}
\usage{
init_tic(commit = TRUE)
}
\arguments{
\item{commit}{`[logical]`\cr
Whether to automatically commit the newly created files.}
}
\description{
Initialize tic CI structure
}
