% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_out_of_tree.R
\name{move_out_of_tree}
\alias{move_out_of_tree}
\title{A utility to move a directory out of the current package tree}
\usage{
move_out_of_tree(path)
}
\arguments{
\item{path}{The directory or file to be moved out of tree.
Must be a subdirectory of the current working directory.}
}
\description{
Useful for revdepcheck results and other large directories or files
that are not needed for the package to function.
The directory will be moved to a sibling directory of the
current package directory.
}
