% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_downgrade}
\alias{renv_downgrade}
\title{Downgrade an renv project to a specific RSPM snapshot}
\usage{
renv_downgrade()
}
\value{
Called for its side-effect.
}
\description{
This functions aims to be used within a "snapshot-centered project
workflow" and can be used to downgrade all packages to an RSPM snapshot listed in
\code{renv.lock}.

While the main purpose is downgrade packages which exist in a higher
version, this function can also be used to restore a clean state of the
project library outside of a downgrade scenario.

Under the hood, it records all packages installed in the \code{renv} project
library and restores these with the RSPM snapshot found in Line 7 of
\code{renv.lock}.
}
\examples{
\dontrun{
renv_downgrade()
}
}
\seealso{
renv_switch_r_version
}
