% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_install_local}
\alias{renv_install_local}
\title{Build a local package and install it into an renv project}
\usage{
renv_install_local(path = ".", quiet = FALSE, ...)
}
\arguments{
\item{path}{\verb{[character]}\cr
The path to the package which should be built and installed.}

\item{quiet}{\verb{[logical]}\cr
Whether to suppress console output.}

\item{...}{\cr
Passed down to \code{pkgbuild::build()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a wrapper around \code{pkgbuild::build()} and \code{renv::install()} to more
easily make local packages available within \pkgrenv projects.

The following steps are performed:
\enumerate{
\item Building the package found at argument \code{path} via \code{pkgbuild::build()}.
\item Moving the built source into the \pkgrenv cache. The cache location is
determined by \code{Sys.getenv("RENV_PATHS_LOCAL")}.
\item Installing the package from the cache location via \code{renv::install()}.
}
}
