% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aviator.R
\name{use_aviator}
\alias{use_aviator}
\title{Install Aviator for a repository}
\usage{
use_aviator()
}
\description{
Aviator is a GitHub app that provides a merge queue.
Compared to GitHub's native merge queue, it does not require branch protection.

This function:
\itemize{
\item installs the Aviator app for a repository (FIXME)
\item creates a default configuraton and pushes it to the main branch
}
}
