% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cmakelists.R
\name{use_cmakelists}
\alias{use_cmakelists}
\title{Initialize CMakeLists configuration}
\usage{
use_cmakelists(project = "NewProject")
}
\arguments{
\item{project}{\verb{[character]}\cr
The name of the R project}
}
\description{
Initializes a \code{CMakeLists.txt} at the current working directory and at the \verb{src/}
folder with the necessary configuration information
}
\examples{
\dontrun{
use_cmakelists("Your_Project_Name")
}
}
