% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upkeep.R
\name{use_cynkra_upkeep_issue}
\alias{use_cynkra_upkeep_issue}
\title{Create an issue containing a checklist of upkeep tasks}
\usage{
use_cynkra_upkeep_issue(repo = NULL)
}
\arguments{
\item{repo}{GitHub repository, e.g. "cynkra/cynkrathis". By default,
the repository associated to the current folder.}
}
\value{
Nothing, used for its side-effect.
}
\description{
Bring your package up to current cynkra development standards.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
use_cynkra_upkeep_issue()
\dontshow{\}) # examplesIf}
}
