% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitpod.R
\name{use_gitpod}
\alias{use_gitpod}
\title{Initialize gitpod configuration}
\usage{
use_gitpod()
}
\description{
Initializes \code{.gitpod.Dockerfile} and \code{.gitpod.yml} at the working directory
with the necessary gitpod configuration information
}
\examples{
\dontrun{
use_gitpod()
}
}
