% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{init_renv}
\alias{init_renv}
\title{Initialize {renv} infrastructure (cynkra way)}
\usage{
init_renv(snapshot_date = NULL, exclude = NULL, convenience_pkgs = FALSE)
}
\arguments{
\item{snapshot_date}{`[integer]`\cr
A valid RSPM snapshot date. By default the "recommended" date from
`get_snapshots()` for the respective R version is chosen.}

\item{exclude}{`[character]`\cr
Packages to exclude from `renv::install()`.
Useful if a package is not available in the available repositories
(check with `getOption("repos"`) to prevent `init_renv()` to fail.
This should only be a temporary workaround - consider making local
packages available in a minicran-like repository.}

\item{convenience_pkgs}{`[logical]`\cr
  Install additional opinionated convenience packages?
  The following packages would be installed:

  - usethis
  - styler
  - gert
  - krlmlr/fledge}
}
\description{
Initializes {renv} setup by setting a predefined RStudio Package
Manager (RSPM) snapshot.
Custom RSPM Snapshots can be configured via `snapshot_date`.
}
\details{
During the process, the latest CRAN version of {renv} will be installed,
regardless of the chose snapshot ID.

The heuristic for setting the correct RSPM binary repo currently only supports
Windows, macOS and Ubuntu 20.04.
}
\examples{
\dontrun{
init_renv()
}
}
