% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{init_renv}
\alias{init_renv}
\title{Initialize {renv} infrastructure (cynkra way)}
\usage{
init_renv(snapshot_date = NULL, exclude = NULL, convenience_pkgs = FALSE)
}
\arguments{
\item{snapshot_date}{\verb{[Date]}\cr
A valid RSPM snapshot date. By default the "recommended" date from
\code{get_snapshots()} for the respective R version is chosen.}

\item{exclude}{\verb{[character]}\cr
Packages to exclude from \code{renv::install()}.
Useful if a package is not available in the available repositories
(check with \verb{getOption("repos"}) to prevent \code{init_renv()} to fail.
This should only be a temporary workaround - consider making local
packages available in a minicran-like repository.}

\item{convenience_pkgs}{\verb{[logical]}\cr
Install additional opinionated convenience packages?
The following packages would be installed:
\itemize{
\item usethis
\item styler
\item gert
\item krlmlr/fledge
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Initializes {renv} setup by setting a predefined RStudio Package
Manager (RSPM) snapshot.
Custom RSPM Snapshots can be configured via \code{snapshot_date}.
}
\details{
During the process, the latest CRAN version of {renv} will be installed,
regardless of the chosen snapshot ID.

The heuristic for setting the correct RSPM binary repo currently only supports
Windows, macOS and Ubuntu 20.04.

The initialization mostly runs in clean vanilla sessions started with
\code{\link[callr:r_vanilla]{callr::r_vanilla()}}.
}
\examples{
\dontrun{
init_renv()
}
}
