% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_install_local}
\alias{renv_install_local}
\title{Build a local package and install it into an renv project}
\usage{
renv_install_local(path = ".")
}
\description{
This is a wrapper around \code{pkgbuild::build()} and
\code{renv::install()} to more easily make local packages available within
\pkg{renv} projects.

The following steps are performed:
\enumerate{
\item Building the package found at argument \code{path} via \code{pkgbuild::build()}.
\item Moving the built source into the \pkg{renv} cache. The cache location is
determined by \code{Sys.getenv("RENV_PATHS_LOCAL")}.
\item Installing the package from the cache location via \code{renv::install()}.
}
}
