% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_switch_r_version}
\alias{renv_switch_r_version}
\title{Switch between R versions in renv projects}
\usage{
renv_switch_r_version(version = NULL, update_packages = TRUE, snapshot = TRUE)
}
\description{
This function switches between R versions in renv projects
which follow the 'cynkra RSPM snapshot' logic.
This is done by replacing the respective entries in \code{renv.lock}.

The following tasks are executed
\itemize{
\item Replace the R Version in \code{renv.lock}
\item Replace the RSPM snapshot in \code{renv.lock} with the one associated with the
selected R Version
\item (optional) execution of \code{renv::update()}
\item (optional) execution of \code{renv::snapshot()}
}
}
\examples{
renv_switch_r_version("4.0.4")
}
\seealso{
get_valid_snapshots
}
