% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_switch_r_version}
\alias{renv_switch_r_version}
\title{Switch between R versions in renv projects}
\usage{
renv_switch_r_version(
  version = NULL,
  update_packages = FALSE,
  snapshot = FALSE
)
}
\arguments{
\item{version}{\verb{[character]}\cr
The R version to upgrade to.}

\item{update_packages}{\verb{[logical]}\cr
Whether to update all packages to the new RSPM snapshot via
\code{renv::update()}.}

\item{snapshot}{\verb{[logical]}\cr
Whether to call \code{renv::snapshot()} after all packages have been updated.}
}
\value{
TRUE (invisibly)
}
\description{
This function switches between R versions in renv projects
which follow the 'cynkra RSPM snapshot' logic.
This is done by replacing the respective entries in \code{renv.lock}.

The following tasks are executed
\itemize{
\item Replace the R Version in \code{renv.lock}
\item Replace the RSPM snapshot in \code{renv.lock} with the one associated with the
selected R Version
\item (optional) execution of \code{renv::update()}
\item (optional) execution of \code{renv::snapshot()}
}
}
\details{
Currently only R version upgrades are supported, i.e. going from R 4.0.3 to
R 4.0.4 (or higher).
}
\examples{
\dontrun{
renv_switch_r_version("4.0.4")
}
}
\seealso{
get_snapshots
}
