% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_switch_r_version}
\alias{renv_switch_r_version}
\title{Switch between R versions in renv projects}
\usage{
renv_switch_r_version(version = NULL)
}
\arguments{
\item{version}{\verb{[character]}\cr
The R version to upgrade to.}
}
\value{
TRUE (invisibly)
}
\description{
This function switches between R versions in renv projects
which follow the 'cynkra RSPM snapshot' logic.
This means that each R version is tied to a specific RSPM snapshot.

The function executes the following tasks:
\itemize{
\item Replace the R Version in \code{renv.lock}.
\item Replace the RSPM snapshot in \code{renv.lock} with the one associated with the
selected R Version.
\item (optional) execution of \code{renv::update()}.
\item (optional) execution of \code{renv::snapshot()}.
}
}
\examples{
\dontrun{
renv_switch_r_version("4.0.4")
}
}
\seealso{
get_snapshots
}
